extract persistence module
