+++ persistence/src/main/kotlin/tw/waterballsa/utopia/persistence/actuator/Extension.kt
@@ -1,0 +1,12 @@
+package tw.waterballsa.utopia.mongo.actuator
+
+import tw.waterballsa.utopia.actuator.PingPongRecord
+import java.time.Instant
+
+fun PingPongRecord.toDocument(): PingPongDocument {
+    return PingPongDocument(name, memo, Instant.now())
+}
+
+fun PingPongDocument.toRecord(): PingPongRecord {
+    return PingPongRecord(name, memo)
+}
